% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{buildModelInputs}
\alias{buildModelInputs}
\title{Builds ProtModel input data}
\usage{
buildModelInputs(
  data,
  form,
  formM,
  protColName,
  peptColName,
  cleanNonIdentifiable = FALSE,
  contrasts
)
}
\arguments{
\item{data}{the Summarized Experiment}

\item{form}{the formula used to define the main intensity linear model}

\item{formM}{the formula used to define the missingness model}

\item{protColName}{A character(1) which designates the name of the column
in \code{data@rowData} that contains the protein identifiers}

\item{peptColName}{A character(1) which designates the name of the column
in \code{data@rowData} that contains the peptide identifiers}

\item{cleanNonIdentifiable}{TRUE if we clean multicollinearity explicitly}
}
\value{
A \code{list()} containing various STAN model input data
}
\description{
This function takes raw data input and builds ProtModel input data.
}
\author{
Philippe Hauchamps
}
