% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{check_all_diagnostics}
\alias{check_all_diagnostics}
\title{Run all diagnostics one at a time}
\usage{
check_all_diagnostics(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Run all diagnostics one at a time
}
