% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{check_div}
\alias{check_div}
\title{Check transitions that ended with a divergence}
\usage{
check_div(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Check transitions that ended with a divergence
}
