% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{check_energy}
\alias{check_energy}
\title{Check energy Bayesian fraction of missing information (E-BFMI)}
\usage{
check_energy(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Check energy Bayesian fraction of missing information (E-BFMI)
}
