% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{check_n_eff}
\alias{check_n_eff}
\title{Check effective sample size per iteration}
\usage{
check_n_eff(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Check effective sample size per iteration
}
