% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{check_rhat}
\alias{check_rhat}
\title{Checks potential scale reduction factors}
\usage{
check_rhat(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Checks potential scale reduction factors
}
