% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{check_treedepth}
\alias{check_treedepth}
\title{Check transitions that ended prematurely due to maximum tree depth limit}
\usage{
check_treedepth(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Check transitions that ended prematurely due to maximum tree depth limit
}
