% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStanDiagnostics.R
\name{checkAllDiagnostics}
\alias{checkAllDiagnostics}
\title{Run all diagnostics one at a time, by model}
\usage{
checkAllDiagnostics(x, displayMessages = TRUE)
}
\arguments{
\item{x}{a \code{ProtModelFitStan} object}
}
\description{
Run all diagnostics one at a time, by model
}
