% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStanDiagnostics.R
\name{checkDiv}
\alias{checkDiv}
\title{Check transitions that ended with a divergence, by model}
\usage{
checkDiv(x, displayMessages = TRUE)
}
\arguments{
\item{x}{a \code{ProtModelFitStan} object}
}
\description{
Check transitions that ended with a divergence, by model
}
