% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStanDiagnostics.R
\name{checkEnergy}
\alias{checkEnergy}
\title{Check energy Bayesian fraction of missing information (E-BFMI), by
model}
\usage{
checkEnergy(x, displayMessages = TRUE)
}
\arguments{
\item{x}{a \code{ProtModelFitStan} object}
}
\description{
Check energy Bayesian fraction of missing information (E-BFMI), by
model
}
