% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStanDiagnostics.R
\name{checkNEff}
\alias{checkNEff}
\title{Check effective sample size per iteration, by model}
\usage{
checkNEff(x, displayMessages = TRUE)
}
\arguments{
\item{x}{a \code{ProtModelFitStan} object}
}
\description{
Check effective sample size per iteration, by model
}
