% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStanDiagnostics.R
\name{checkRhat}
\alias{checkRhat}
\title{Checks potential scale reduction factors, by model}
\usage{
checkRhat(x, displayMessages = TRUE)
}
\arguments{
\item{x}{a \code{ProtModelFitStan} object}
}
\description{
Checks potential scale reduction factors, by model
}
