% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStanDiagnostics.R
\name{checkTreeDepth}
\alias{checkTreeDepth}
\title{Check transitions that ended prematurely due to maximum tree depth
limit, by model}
\usage{
checkTreeDepth(x, displayMessages = TRUE)
}
\arguments{
\item{x}{a \code{ProtModelFitStan} object}
}
\description{
Check transitions that ended prematurely due to maximum tree depth
limit, by model
}
