% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{get_total_elapsed_time}
\alias{get_total_elapsed_time}
\title{Returns total elapsed time that was needed to run stan}
\usage{
get_total_elapsed_time(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Returns total elapsed time that was needed to run stan
}
