% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStanDiagnostics.R
\name{getElapsedTime}
\alias{getElapsedTime}
\title{Returns total elapsed time that was needed to run a ProtModelFitStan engine}
\usage{
getElapsedTime(x)
}
\arguments{
\item{x}{a \code{ProtModelFitStan} object}
}
\description{
Returns total elapsed time that was needed to run a ProtModelFitStan engine
}
