% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{Create a ProtModel object}
\usage{
makeModel(
  modelName,
  data,
  formula,
  formulaM,
  protColName,
  peptColName,
  cleanNonIdentifiable = FALSE,
  refIntensity = 0,
  contrasts = list()
)
}
\arguments{
\item{modelName}{A \code{character(1)} containing the name of the model.}

\item{data}{A \code{SummarizedExperiment} object.}

\item{formula}{A \code{formula} used to define the main intensity linear model}

\item{formulaM}{A \code{formula} used to define the missingness model}

\item{protColName}{A character(1) which designates the name of the column
in \code{data@rowData} that contains the protein identifiers}

\item{peptColName}{A character(1) which designates the name of the column
in \code{data@rowData} that contains the peptide identifiers}

\item{cleanNonIdentifiable}{A boolean indicating whether the engine should
try to remove regressors for those protein models that are not identifiable}
}
\value{
An initialized \code{StanModel} object.
}
\description{
This function takes a data set and a formula and initiates a
\code{ProtModel}.
}
\author{
Laurent Gatto
}
