% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfitDiagnostics.R
\name{partition_div}
\alias{partition_div}
\title{Returns parameter arrays separated into divergent and non-divergent transitions}
\usage{
partition_div(object)
}
\arguments{
\item{object}{a \code{stanfit} object}
}
\description{
Returns parameter arrays separated into divergent and non-divergent transitions
}
