% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModel.R
\name{proteinInfo}
\alias{proteinInfo}
\title{Get data aggregation of per protein information}
\usage{
proteinInfo(object, proteins = NULL)
}
\arguments{
\item{object}{a \code{ProtModel} object}
}
\description{
Get data aggregation of per protein information
}
