% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModel.R
\docType{class}
\name{ProtModel-class}
\alias{ProtModel-class}
\alias{.ProtModel}
\alias{ProtModel-class,}
\alias{ProtModel}
\title{ProtModel class}
\usage{
ProtModel(
  .modelName = "ProtModelName",
  .data = SummarizedExperiment(),
  .formula = formula(),
  .formulaM = formula(),
  .InputData = list(),
  .proteinInfo = data.frame(),
  .paramNames = list()
)
}
\arguments{
\item{.modelName}{A \code{character(1)} containing the name of the model.}

\item{.data}{A \code{SummarizedExperiment} object.}

\item{.formula}{A \code{formula} defining the main intensity linear model.}

\item{.formulaM}{A \code{formula} defining the missingness linear model.}

\item{.InputData}{A \code{list} of input data that will be passed to the model
engine upon a run.}

\item{.proteinInfo}{A \code{data.frame} of per protein information (nb peptides,
nb observations, \%missing data, ...).}

\item{.paramNames}{A \code{list} of model parameter names by category.}
}
\value{
An object of class \code{ProtModel}
}
\description{
This class encapsulates a model fit for proteomics data analysis.

Object can be created manually with \code{ProtModel()} or using the
\code{\link[=makeModel]{makeModel()}} function.
}
\section{Slots}{

\describe{
\item{\code{modelName}}{A \code{character(1)} containing the name of the model.}

\item{\code{data}}{The proteomics data as a \code{SummarizedExperiment}.}

\item{\code{formula}}{The formula defining the main intensity linear model.}

\item{\code{formulaM}}{The formula defining the missingness linear model.}

\item{\code{run}}{\code{numeric(1)} where 0 denotes an initialised model that
hasn't been run. Any value > 1 indicates that the parameters
have been estimated.}

\item{\code{InputData}}{A \code{list()} of input data that will be passed to the model
engine upon a run.}

\item{\code{proteinInfo}}{A \code{data.frame} of per protein information (nb peptides,
nb observations, \%missing data, ...)}

\item{\code{paramNames}}{A \code{list} of model parameter names by category, generated from
the parsing of the formulas}

\item{\code{ModelFit}}{A \code{ProtModelFit} object which is created at run time}

\item{\code{lastModelSummary}}{A \code{list} containing the args and the summary results
of the last computed summary}
}}

\examples{
ProtModel()
}
