% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFit.R, R/ProtModelFitMFVB.R
\docType{class}
\name{ProtModelFit-class}
\alias{ProtModelFit-class}
\alias{.ProtModelFit}
\alias{ProtModelFit-class,}
\alias{ProtModelFit}
\alias{ProtModelFitMFVB}
\title{ProtModelFit class}
\usage{
ProtModelFit(
  .ModelSpecPars = ProtModelSpecPars(),
  .algoArgs = list(),
  .fullStorage = FALSE
)

ProtModelFitMFVB(
  .ModelSpecPars = ProtModelSpecPars(),
  .algoArgs = list(),
  .fullStorage = FALSE,
  .fitResults = list(),
  .paramDistr = list()
)
}
\arguments{
\item{.ModelSpecPars}{A \code{StanModelSpecPars} object containing the model
specification parameters.}

\item{.algoArgs}{A \code{list} object containing algorithm tuning parameters}

\item{.fullStorage}{A \code{logical} indicating whether full storage of parameters
information is done after run}

\item{.fitResults}{A \code{list()} of results related to the fit procedure}

\item{.paramDistr}{A \code{list()} containing the per parameter probability
distributions generated by the MFVB algorithm.}
}
\value{
An object of class \code{ProtModelFit}

An object of class \code{ProtModelFitMFVB}
}
\description{
This class encapsulates a generic model fit for proteomics data analysis.

Object is created internally when runModel() is called on a ProtModel object
and stored as a slot in this object
}
\section{Slots}{

\describe{
\item{\code{ModelSpecPars}}{A \code{ProtModelSpecPars} object containing STAN model
specification parameters.}

\item{\code{algoArgs}}{A \code{list()} containing tuning arguments for the ModelFit engine}

\item{\code{fullStorage}}{A 'logical' indicating whether full storage of the
parameters data will be done after engine run}
}}

