% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitMFVB.R
\docType{class}
\name{ProtModelFitMFVB-class}
\alias{ProtModelFitMFVB-class}
\alias{.ProtModelFitMFVB}
\alias{ProtModelFitMFVB-class,}
\alias{ProtModelFitMFVB}
\title{ProtModelFitMFVB class}
\description{
This class encapsulates a MFVB model fit for proteomics data analysis.

Object is created internally when runModel() is called on a StanProt object
and stored as a slot in this object
}
\section{Slots}{

\describe{
\item{\code{fitResults}}{A \code{list()} of results related to the fit procedure}

\item{\code{distrMap}}{A \code{list()} containing the static mapping to the various
distributions functions}

\item{\code{paramDistr}}{A \code{list()} containing the per parameter probability
distributions info generated by the MFVB algorithm.}
}}

