% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelFitStan.R
\docType{class}
\name{ProtModelFitStan-class}
\alias{ProtModelFitStan-class}
\alias{.ProtModelFitStan}
\alias{ProtModelFitStan-class,}
\alias{ProtModelFitStan}
\title{ProtModelFitStan class}
\usage{
ProtModelFitStan(
  .modelScript = NULL,
  .runMode = "sampling",
  .ModelSpecPars = StanModelSpecPars(),
  .algoArgs = list(),
  .recompileModelEachTime = FALSE,
  .fullStorage = FALSE
)
}
\arguments{
\item{.modelScript}{A \code{character(1)} designating the model script to be used}

\item{.runMode}{A \code{character(1)} specifying the run mode of the stan fit object
can be 'sampling', 'MFVB', or 'MAP'}

\item{.ModelSpecPars}{A \code{StanModelSpecPars} object containing the model
specification parameters.}

\item{.algoArgs}{A \code{list} object containing algorithm tuning parameters}

\item{.recompileModelEachTime}{A \code{logical} stating whether stan model should
be recompiled anyway, even when compiled object is present from disk.}

\item{.fullStorage}{A \code{logical} indicating whether full storage will
be done after calculation}
}
\value{
An object of class \code{ProtModelFitStan}
}
\description{
This class encapsulates a stan model fit for proteomics data analysis.

Object is created internally when runModel() is called on a StanProt object
and stored as a slot in this object
}
\section{Slots}{

\describe{
\item{\code{modelScript}}{A \code{character(1)} designating the model script to be used}

\item{\code{runMode}}{A \code{character(1)} specifying the run mode of the stan fit object
can be 'sampling', 'MFVB', or 'MAP'}

\item{\code{recompileModelEachTime}}{A \code{logical} stating whether stan model should
be recompiled anyway, even when compiled object is present from disk
This is done as a work-around to issue observed when using stan together
with BiocParalel (MulticoreParam)}

\item{\code{stanResList}}{A \code{list} of results which are generated by STAN at run time.
Multi-element list is created when running the Stan model protein by protein.
A list element containing only one or two entries which can be :
\itemize{
\item \code{Stanfit} : a Stanfit object if runMode == "sampling" or "MFVB" and
fullStorage is TRUE
\item \code{StanMapRes} : the stan generated object when runMode == "MAP" whatever
the value of fullStorage (TRUE or FALSE)
\item \code{StanSummary} : a summary table with mean, sd, and standard quantiles
by parameter if fullStorage is FALSE (all run modes, except MAP)
\item \code{StanDiagnostics} : a list containing diagnostic elements if fullStorage
is FALSE (all run modes, except MAP)
}}

\item{\code{protIndexList}}{A \code{list} of indexes variables, only used internally to
access and converts the different stanfit objects parameter names, when the
Stan model has been run protein by protein.}
}}

