% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtModelSpecPars.R
\docType{class}
\name{ProtModelSpecPars-class}
\alias{ProtModelSpecPars-class}
\alias{.ProtModelSpecPars}
\alias{ProtModelSpecPars-class,}
\alias{ProtModelSpecPars}
\title{ProtModelSpecPars class}
\description{
This class encapsulates a set of model specifications for a ProtModelFit.

Object can be created manually with \code{ProtModelSpecPars()}.
}
\section{Slots}{

\describe{
\item{\code{priorIntensityMean}}{The mean of the (base) intensity prior distribution.}

\item{\code{priorFoldChangeMean}}{The mean of the log2Fold changes prior distribution.}

\item{\code{priorIntensitySd}}{The std dev of the (base) intensity prior distribution.}

\item{\code{priorFoldChangeSd}}{The std dev of the log2Fold changes prior distribution.}

\item{\code{priorGammaInterceptMean}}{The prior mean of the intercept coefficient
in the missingness model}

\item{\code{priorGammaTermsMean}}{The prior mean of the other MAR coefficients
in the missingness model. Either there is only one mean, and then it is
interpreted as a common prior mean for all MAR gamma terms (except intercept),
or it is a vector.}

\item{\code{priorGammaInterceptSd}}{The prior std dev of the intercept coefficient
in the missingness model}

\item{\code{priorGammaTermsSd}}{The prior std dev of the other MAR coefficients
in the missingness model. Either there is only one mean, and then it is
interpreted as a common prior mean for all MAR gamma terms (except intercept),
or it is a vector.}

\item{\code{priorGammaYMean}}{The prior mean of the NMAR gamma coefficients in the
missingness model}

\item{\code{priorGammaYSd}}{the prior std dev of the NMAR gamma coefficient in the
missingness model}

\item{\code{priorNuEpsilon}}{the Nu coefficient of the prior inv chi2 distribution
for the intensity model errors variances (used in non hierarchical set-up
only)}

\item{\code{priorScaleEpsilon}}{the Scale coefficient of the prior inv chi2
distribution for the intensity model errors variances (used in non
hierarchical set-up only)}

\item{\code{priorNuRdm}}{the Nu coefficient of the prior inv chi2 distribution for
the intensity model random effects variances (used in non hierarchical
set-up only)}

\item{\code{priorScaleRdm}}{the Scale coefficient of the prior inv chi2 distribution
for the intensity model random effects variances (used in non hierarchical
set-up only)}

\item{\code{t_likelihood}}{A \code{logical} indicating whether we use a student-t
distribution for the peptide intensities (\code{TRUE}) or normal (\code{FALSE})}

\item{\code{t_fixedDF}}{A \code{numeric} designing the degrees of freedom of the student
intensity data distribution (if used)}

\item{\code{modelMiss}}{A \code{logical} indicating whether we model missingness or not.}

\item{\code{NMAR}}{A \code{logical} indicating whether we model a NMAR missingness
component or not.}

\item{\code{sigmaHierarch}}{A \code{logical} indicating whether we use a hierarchical
set up for the error residual variances, or not.}

\item{\code{sigmaRdmHierarch}}{A \code{logical} indicating whether, for the random
effects (if any), we use a hierarchical set-up for the random effect
variances, or not.}

\item{\code{gammaHierarch}}{A \code{logical} indicating whether we use a hierarchical
set up for the missingness model coefficients (gamma & gammY), or not.}
}}

\examples{
ProtModelSpecPars()

}
