% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Run a proteomics model}
\usage{
runModel(
  x,
  modelScript = NULL,
  providerPackage = c("stan", "native"),
  runMode = c("sampling", "MFVB", "MAP"),
  modelSpecPars = NULL,
  algoArgs = NULL,
  initialValues = NULL,
  fullStorage = TRUE,
  forceRecompile = FALSE
)
}
\arguments{
\item{x}{A \code{ProtModel} object as produced by \code{\link[=makeModel]{makeModel()}}.}

\item{modelScript}{\code{character(1)} provide the name without extension
to a stan model file. The function will check if either
\code{stanmodel.rds} or \code{stanmodel.stan} are available an use the
more recent one. Only used in 'stan' mode}

\item{modelSpecPars}{an optional \code{ProtModelSpecPars} of parameters tuning
the specification of the proteomics model.
If not provided, will use previously set \code{ModelSpecPars} within \code{x}}

\item{algoArgs}{an optional list of tuning arguments to be provided to the
engine. If not provided, will use previously set \code{algoArgs}
within \code{x}}

\item{mode}{\code{character(1)} can be either 'stan' or 'native'}

\item{algo}{\code{character(1)} can be :
\itemize{
\item 'sampling' for exact inference,
\item 'mfvb' for approximate inference using mean field variational bayes,
\item 'MAP' for maximum a posteriori, i.e. mode finding using numerical
optimization.
}}
}
\value{
A updated \code{ProtModel} object.
}
\description{
Take an initialised \code{ProtModel} object, runs the inference engine and
returns an updated \code{ProtModel} object containing the estimated coefficients.
}
\author{
Philippe Hauchamps
}
