% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProtModelEngine.ProtModelFitMFVB.R
\name{runProtModelEngine.ProtModelFitMFVB}
\alias{runProtModelEngine.ProtModelFitMFVB}
\title{Run a Bayes model using variational approximation}
\usage{
runProtModelEngine.ProtModelFitMFVB(x, InputData, initialValues)
}
\arguments{
\item{x}{A \code{ProtModelFitMFVB} object}
}
\value{
A updated \code{ProtModelFitMFVB} object.
}
\description{
Take an initialised \code{ProtModelFitMFVB} object, runs MFVB approximation
algorithm and returns an updated \code{ProtModelFitMFVB} object containing the
estimated parameters distributions
}
\author{
Philippe Hauchamps
}
