% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{updateParamNames}
\alias{updateParamNames}
\title{Update parameter names and mappings.}
\usage{
updateParamNames(x)
}
\arguments{
\item{x}{ProtModel object}
}
\value{
\code{x} with an updated model parameter names mappings
}
\description{
This function Updates the model parameter names of a ProtModel object,
organized by category, as a result of the formulas parsing.
}
\author{
Philippe Hauchamps
}
